+++ apps/fluent-tester/src/TestComponents/RadioGroupExperimental/index.ts
@@ -1,0 +1,2 @@
+export * from './RadioGroupExperimentalTest';
+export * from './consts';