+++ packages/experimental/Badge/src/BadgeColorTokens.ts
@@ -6,41 +6,14 @@ import { getFilledColorProps, getOutlineColorProps, getTintColorProps, getGhostC
 export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
   ({
     filled: {
-      ...getFilledColorProps(
-        {
-          backgroundColor: t.colors.brandBackgroundStatic,
-        },
-        t,
-      ),
       borderColor: 'transparent',
     },
     outline: {
-      ...getOutlineColorProps({ color: t.colors.brandForeground1, colorDark: t.colors.brandForeground1 }, t),
       backgroundColor: t.colors.transparentBackground,
-    },
-    tint: {
-      ...getTintColorProps(
-        {
-          backgroundColor: globalTokens.color.brand.tint60,
-          color: t.colors.brandForeground1,
-          borderColor: t.colors.brandStroke2,
-          backgroundColorDark: globalTokens.color.outlook.shade40,
-          colorDark: globalTokens.color.brand.tint30,
-          borderColorDark: globalTokens.color.outlook.shade40,
-        },
-        t,
-      ),
     },
     ghost: {
       backgroundColor: t.colors.transparentBackground,
       borderColor: t.colors.transparentStroke,
-      ...getGhostColorProps(
-        {
-          color: t.colors.brandForeground1,
-          colorDark: t.colors.brandForeground1,
-        },
-        t,
-      ),
     },
     brand: {
       filled: {
@@ -52,17 +25,14 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: t.colors.brandForeground1, colorDark: t.colors.brandForeground1 }, t),
+        ...getOutlineColorProps({ color: t.colors.brandForeground1 }, t),
       },
       tint: {
         ...getTintColorProps(
           {
-            backgroundColor: globalTokens.color.brand.tint60,
-            color: t.colors.brandForeground1,
+            backgroundColor: t.colors.brandBackground2,
+            color: t.colors.brandForeground2,
             borderColor: t.colors.brandStroke2,
-            backgroundColorDark: globalTokens.color.outlook.shade40,
-            colorDark: globalTokens.color.brand.tint30,
-            borderColorDark: globalTokens.color.outlook.shade40,
           },
           t,
         ),
@@ -71,7 +41,6 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
         ...getGhostColorProps(
           {
             color: t.colors.brandForeground1,
-            colorDark: t.colors.brandForeground1,
           },
           t,
         ),
@@ -152,7 +121,7 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
         ...getFilledColorProps(
           {
             backgroundColor: globalTokens.color.yellow.primary,
-            color: globalTokens.color.grey[14],
+            color: globalTokens.color.grey[14], // It should be neutralForegroundStatic1. It's hardcoded because the token doesn't exist right now
           },
           t,
         ),
@@ -222,22 +191,21 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       filled: {
         ...getFilledColorProps(
           {
-            backgroundColor: globalTokens.color.grey[14],
-            backgroundColorDark: globalTokens.color.white,
-            colorDark: globalTokens.color.grey[14],
+            backgroundColor: t.colors.neutralForeground1,
+            color: t.colors.neutralBackground1,
           },
           t,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: globalTokens.color.grey[14], colorDark: globalTokens.color.white }, t),
+        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStrokeAccessible }, t),
       },
       tint: {
         ...getTintColorProps(
           {
-            backgroundColor: globalTokens.color.grey[38],
-            color: t.colors.neutralForegroundOnBrand,
-            borderColor: t.colors.transparentBackground,
+            backgroundColor: t.colors.neutralForeground3,
+            color: t.colors.neutralBackground1,
+            borderColor: t.colors.neutralStrokeAccessible,
             backgroundColorDark: globalTokens.color.grey[68],
             colorDark: globalTokens.color.grey[16],
             borderColorDark: globalTokens.color.grey[68],
@@ -248,8 +216,7 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       ghost: {
         ...getGhostColorProps(
           {
-            color: globalTokens.color.grey[14],
-            colorDark: globalTokens.color.white,
+            color: t.colors.neutralForeground1,
           },
           t,
         ),
@@ -259,32 +226,23 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       filled: {
         ...getFilledColorProps(
           {
-            backgroundColor: globalTokens.color.grey[92],
-            color: globalTokens.color.grey[38],
-            backgroundColorDark: t.colors.black,
-            colorDark: globalTokens.color.grey[68],
+            backgroundColor: t.colors.neutralBackground5,
+            color: t.colors.neutralForeground3,
             hcBackground: t.colors.neutralBackground3,
             hcColor: t.colors.brandForeground1,
-            hcBorderColor: t.colors.brandForeground1,
           },
           t,
         ),
       },
       outline: {
-        ...getOutlineColorProps(
-          { color: globalTokens.color.grey[92], colorDark: globalTokens.color.grey[68], borderColorDark: globalTokens.color.grey[32] },
-          t,
-        ),
+        ...getOutlineColorProps({ color: t.colors.neutralForeground3, borderColor: t.colors.neutralStroke2 }, t),
       },
       tint: {
         ...getTintColorProps(
           {
-            backgroundColor: globalTokens.color.grey[94],
-            color: globalTokens.color.grey[38],
-            borderColor: globalTokens.color.grey[92],
-            backgroundColorDark: globalTokens.color.grey[8],
-            colorDark: globalTokens.color.grey[68],
-            borderColorDark: globalTokens.color.grey[32],
+            backgroundColor: t.colors.neutralBackground4,
+            color: t.colors.neutralForeground3,
+            borderColor: t.colors.neutralStroke2,
           },
           t,
         ),
@@ -292,8 +250,7 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       ghost: {
         ...getGhostColorProps(
           {
-            color: globalTokens.color.grey[92],
-            colorDark: globalTokens.color.grey[68],
+            color: t.colors.neutralForeground3,
           },
           t,
         ),
@@ -303,29 +260,23 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       filled: {
         ...getFilledColorProps(
           {
-            backgroundColor: t.colors.white,
-            color: globalTokens.color.grey[14],
-            backgroundColorDark: globalTokens.color.grey[16],
-            colorDark: t.colors.neutralForegroundOnBrand,
+            backgroundColor: t.colors.neutralBackground1,
+            color: t.colors.neutralForeground1,
             hcBackground: t.colors.neutralBackground3,
             hcColor: t.colors.brandForeground1,
-            hcBorderColor: t.colors.brandForeground1,
           },
           t,
         ),
       },
       outline: {
-        ...getOutlineColorProps({ color: globalTokens.color.white, colorDark: globalTokens.color.white }, t),
+        ...getOutlineColorProps({ color: t.colors.neutralForegroundOnBrand }, t),
       },
       tint: {
         ...getTintColorProps(
           {
-            backgroundColor: globalTokens.color.white,
-            color: globalTokens.color.grey[38],
-            borderColor: globalTokens.color.grey[88],
-            backgroundColorDark: globalTokens.color.grey[29],
-            colorDark: globalTokens.color.grey[68],
-            borderColorDark: globalTokens.color.grey[32],
+            backgroundColor: t.colors.neutralBackground1,
+            color: t.colors.neutralForeground3,
+            borderColor: t.colors.neutralStroke2,
           },
           t,
         ),
@@ -333,8 +284,7 @@ export const defaultBadgeColorTokens: TokenSettings<BadgeTokens> = (t: Theme) =>
       ghost: {
         ...getGhostColorProps(
           {
-            color: globalTokens.color.white,
-            colorDark: globalTokens.color.white,
+            color: t.colors.neutralForegroundOnBrand,
           },
           t,
         ),